+++ CHANGE_LOG.md
@@ -8,3 +8,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Create a plug-in which corrects the scope of compile dependencies in generated POM files
 - Create a plug-in which adds standard behavior for sources and javadoc jars
 - Create a plug-in which applies a configuration for specifying published information about a project
+- Create a plug-in which adds standard behavior for sources/JavaDoc artifacts in MavenPublications